This is f64 now

+++ src/toggle_project.rs
@@ -11,7 +11,7 @@ pub fn toggle(scrolls: &mut Vec<f64>) {
   if hash != "" {
     let selector = &format!(".projects .project.{}", hash);
     if document().query_selector(selector).unwrap().is_some() {
-      let top: f64 = Some(window().page_y_offset() as f64)
+      let top: f64 = Some(window().page_y_offset())
         .unwrap_or(document().document_element().unwrap().scroll_top());
       body.set_attribute("data-project", &hash).unwrap();
       scrolls.push(top);
